back away from inline annotations, we can add them at any later time

+++ rtx_core/src/mouth.rs
@@ -393,7 +393,7 @@ impl Mouth {
         if !two_hex {
           // OR ^^ followed by a SINGLE Control char type code???
           let c = self.chars[self.colno + 1];
-          let cn = c as i64;
+          let cn = c as i16;
 
           ch = (cn + if cn >= 64 { -64 } else { 64 }) as u8 as char;
           self.splice(self.colno - 1..self.colno + 2, &[ch]);
